requirements.txt
fastapi
uvicorn
pymongo
passlib
pyjwt
python-dotenv
bcrypt